[AS] merge conflicts
