+++ src/test/java/com/adyen/mirakl/cucumber/stepdefs/MiraklAdyenSteps.java
@@ -48,70 +48,4 @@ public class MiraklAdyenSteps extends StepDefsHelper {
         MiraklCreatedShops shops = miraklShopApi.createBusinessShopWithMissingUboInfo(miraklMarketplacePlatformOperatorApiClient, cucumberTable, legalEntity);
         world.miraklShop = retrieveCreatedShop(shops);
     }
-
-    @Given("^a new (.*) shop has been created in Mirakl with invalid data$")
-    public void aNewBusinessShopHasBeenCreatedInMiraklWithInvalidData(String legalEntity, DataTable table) {
-        List<Map<String, String>> cucumberTable = table.getTableConverter().toMaps(table, String.class, String.class);
-        MiraklCreatedShops shops = miraklShopApi.createBusinessShopWithMissingUboInfo(miraklMarketplacePlatformOperatorApiClient, cucumberTable, legalEntity);
-        world.miraklShop = retrieveCreatedShop(shops);
-    }
-
-    @And("^(?:the previous BankAccountDetail will be removed|a notification will be sent in relation to the balance change)$")
-    public void thePreviousBankAccountDetailWillBeRemoved(DataTable table) {
-        List<Map<String, String>> cucumberTable = table.getTableConverter().toMaps(table, String.class, String.class);
-        String eventType = cucumberTable.get(0).get("eventType");
-        String reason = cucumberTable.get(0).get("reason");
-
-        await().untilAsserted(() -> {
-            DocumentContext adyenNotificationBody = JsonPath.parse(retrieveAdyenNotificationBody(eventType, world.miraklShop.getId()));
-            Assertions.assertThat(adyenNotificationBody.read("content.reason").toString())
-                .contains(reason);
-        });
-    }
-
-    @Then("^a remedial email will be sent for each ubo$")
-    public void aRemedialEmailWillBeSentForEachUbo(String title) throws Throwable {
-
-        GetAccountHolderResponse accountHolder = getGetAccountHolderResponse(world.miraklShop);
-
-        List<String> uboEmails = accountHolder.getAccountHolderDetails().getBusinessDetails().getShareholders().stream()
-            .map(ShareholderContact::getEmail)
-            .collect(Collectors.toList());
-
-        await().untilAsserted(() -> {
-                ResponseBody responseBody = RestAssured.get(mailTrapConfiguration.mailTrapEndPoint()).thenReturn().body();
-                List<Map<String, Object>> emailLists = responseBody.jsonPath().get();
-
-                List<String> htmlBody = new LinkedList<>();
-
-                Assertions.assertThat(emailLists.size()).isGreaterThan(0);
-
-                boolean foundEmail = emailLists.stream()
-                    .anyMatch(map -> map.get("to_email").equals(uboEmails.iterator().next()));
-
-                Assertions.assertThat(foundEmail).isTrue();
-
-                for (String uboEmail : uboEmails) {
-                    emailLists.stream()
-                        .filter(map -> map.get("to_email").equals(uboEmail))
-                        .findAny()
-                        .ifPresent(map -> htmlBody.add(map.get("html_body").toString()));
-                }
-
-                Assertions.assertThat(htmlBody).isNotEmpty();
-
-                for (String body : htmlBody) {
-                    Document parsedBody = Jsoup.parse(body);
-                    Assertions
-                        .assertThat(parsedBody.body().text())
-                        .contains(world.miraklShop.getId());
-
-                    Assertions.assertThat(parsedBody.title()).isEqualTo(title);
-                }
-            }
-        );
-    }
-
-
-
 }
